requirements.txt
py-cord==2.0.0rc1
python-dotenv==0.20.0
pyyaml==6.0